ids changed to int

+++ src/projects/universityAssistant/Services/UniversityService/Core/UniversityService.Domain/Entities/CountryCulture.cs
@@ -4,8 +4,8 @@ namespace UniversityService.Domain.Entities;
 
 public class CountryCulture : Entity
 {
-    public long Id { get; set; }
-    public long CountryId { get; set; }
+    public int Id { get; set; }
+    public int CountryId { get; set; }
     public string Name { get; set; }
     public string Culture { get; set; }
     public virtual Country Country { get; set; }
